backend/requirements.txt
fastapi
uvicorn
groq
python-dotenv
python-multipart
httpx
pydantic